Gateway implementation

+++ food-delivery-user-service-impl/src/main/java/com/example/food/delivery/CustomerAddressServiceImpl.java
@@ -28,7 +28,6 @@ public class CustomerAddressServiceImpl implements CustomerAddressService {
 
     public synchronized ResponseEntity<BaseResponse<?>> addAddress(CustomerAddressRequest custAddressRequest, String custEmail) {
         try {
-            isValidEmail(custEmail);
             if (!customerRepository.existsByCustEmail(custEmail)) {
                 throw new UserManagementExceptions.UserNotFoundException("User with this email not found");
             }
@@ -50,11 +49,14 @@ public class CustomerAddressServiceImpl implements CustomerAddressService {
         return ResponseEntity.ok(response);
     }
 
-    public synchronized ResponseEntity<BaseResponse<?>> deleteCustomerAddress(int addressId) {
+    public synchronized ResponseEntity<BaseResponse<?>> deleteCustomerAddress(int addressId, String custEmail) {
         try {
             CustomerAddress custAddress = custAddressRepository.findById(addressId).orElse(null);
             if (custAddress == null) {
                 throw new UserManagementExceptions.UserNotFoundException("Address Not Found");
+            }
+            if(!custAddress.getCustEmail().equals(custEmail)) {
+                throw new UserManagementExceptions.AddressNotFoundException("Address Id not matches with your email");
             }
             custAddressRepository.deleteById(addressId);
             response = new BaseResponse<>(true, ResponseStatus.SUCCESS.getStatus(), null, "Address removed Successfully");
@@ -66,13 +68,9 @@ public class CustomerAddressServiceImpl implements CustomerAddressService {
 
     public synchronized ResponseEntity<BaseResponse<?>> getCustomerAddress(String customerEmail) {
         try {
-            isValidEmail(customerEmail);
-            if (!customerRepository.existsByCustEmail(customerEmail)) {
-                throw new UserManagementExceptions.UserNotFoundException("User with this email not found");
-            }
             List<CustomerAddress> custAddress = custAddressRepository.findByCustEmail(customerEmail);
             if (custAddress == null || custAddress.isEmpty()) {
-                throw new UserManagementExceptions.UserNotFoundException("No Address Registered");
+                throw new UserManagementExceptions.UserNotFoundException("No Address Registered. Add new Address!");
             }
             response = new BaseResponse<>(true, ResponseStatus.SUCCESS.getStatus(), null, custAddress);
         } catch (Exception e) {
@@ -83,10 +81,6 @@ public class CustomerAddressServiceImpl implements CustomerAddressService {
 
     public synchronized ResponseEntity<BaseResponse<?>> getAddressDetail(String customerEmail, int addressId) {
         try {
-            isValidEmail(customerEmail);
-            if (!customerRepository.existsByCustEmail(customerEmail)) {
-                throw new UserManagementExceptions.AddressNotFoundException("User with this email not found");
-            }
             Optional<CustomerAddress> optCustAddress = custAddressRepository.findById(addressId);
             if (!optCustAddress.isPresent()) {
                 throw new UserManagementExceptions.AddressNotFoundException("No Address Found");
@@ -111,12 +105,4 @@ public class CustomerAddressServiceImpl implements CustomerAddressService {
         return ResponseEntity.ok(response);
     }
 
-    public void isValidEmail(String email) {
-        String regex = "^(.+)@(.+)$";
-        Pattern pattern = Pattern.compile(regex);
-        Matcher matcher = pattern.matcher(email);
-        if(!matcher.matches()) {
-            throw new UserManagementExceptions.InvalidInputException("Oser not found on Id " + email);
-        }
-    }
 }
